Create the first structure of nostr application
This commit is a huge one. It had a lot of feature
and design the whole project. Documentation, tests,
notes, articles, introduction page and main codes
have been added.

All required dependencies have been added, thoas,
gun, and cowboy. Those versions have been fixed.

The current client implementation can use a websocket
to fetch the event from a relay, the connection can
also be closed.

Some functions and modules have been created to permit
to encode and decode payloads from/to a relay/client.
This is not correctly done, yet.

A part of nip/01 is already implemented but requires
a better structure, better testing and better
documentation as well.

The parameters for rebar3 and the tools like compilers
have been created to ensure a good quality for the
produced code.

A Makefile can be used to export the notes/articles in
pdf, epub, plaintext or html formats.