So a short break.  I created the first relay_store handler, to store data in mnesia backend. The idea is to create a behavior with the commands to insert, delete, update and select information from the database. Not ACID... CRUD! :')
Anyway, the goal is to offer some facilities to create those handlers and
extend easily the relay. Inserting an event or a subscription should be
easy on mnesia but also on any kind of database like postgresql.

I think this module will be updated soon with more specification. At least,
it works to insert data in the database (events). The subscription will also
be inserted into the database but the procedure will be a bit different...