So yesterday, we were working on relay store and a mnesia backend. let's continue.
Found a bug in the encoder. Will fix it right now. Better!

Okay, we corrected a bug in the encoder and started to define a new table called
subscriptions. This table will contain a list of subscription identifier (generated
by the clients), the client connected and a reference to a filter (probably another
table). The identifier should be a tuble {ClientIdentifier, Connection}.