--[ SHORTCUT_23/04 ]
We will probably work on another part of the code. We will finish
NIP/02 and merge it. At this time, the relay is doing to much things
and it was not planned at all. Let's work on NIP/02 and the client
tonight!

The problem we had was related related to the way we should receive
the answers from the relay... We will try to deal with that or switch
to another issue.

We already have a great implementation on the client, but we lack it
on the server side. I think it will be done another day. We will switch
to another issue.

----

Let's add the support for the key store. The idea is to store it somewhere
on the disk. To do that we can use many different methods. Which one is the
best? Well... At first, we will just export the key on a file present in the
filesystem. We will improve it later.

Okay, a nostr_client_key module has been created, it is currently not started
but the idea is to have one process per key. This process will be in charge
to save/update/export the private information stored for one user identified
by a name. At first, the private keys (and other informations) will not
be encrypted. It seems a bit crazy, but we need to start with something simple.

This information should be stored outside of the database because it should
be easy to extract and easy to export.

--[ SHORTCUT_23/05 ]

What kind of messages we need to support there? Well. We need to:

 1. (cast) generate a new key
 2. (cast) store the key on the disk
 3. (cast) load the key from the disk (if present)
 4. (call) export it in different formats

The main function created there was the initialization part of
nostr_client_key.

--[ SHORTCUT_23/05 ]

finalizing the initialization part of the nostr_client_key module
using gen_server. This shortcut implemented a way to generate,
initialize, and store a key. More tests are required though.

So, it's over for tonight! 3 shortcuts, one module implemented. It
looks good :)

See you for the next ones ;)